+++ SBO-Lib/ChangeLog.md
@@ -8,6 +8,8 @@
   * sbocheck: Unless in CLASSIC mode, report any SlackBuilds orphaned upstream.
     * Thanks to J. Milgram for the proposal.
   * sbofind, sbotool: Show a removal warning when orphaned builds are displayed.
+  * Util.pm: Allow SBO_HOME/distfiles to be a symlink to an existing directory.
+  * sboclean: Verify a proper SBO_HOME before trying to clean distfiles.
 
 #### 4.2.1 - 2026-07-23
   * sbofind: Show maintainer names unless in CLASSIC mode; search names with --maintainers.